+++ wallet-backend/src/service/voting/crud.rs
@@ -1,11 +1,11 @@
 use crate::repository::choice::model::Choice;
 use crate::repository::voting::model::Voting;
 use crate::repository::voting_config::model::VotingConfig;
-use crate::repository::voting_config::types::ProposerConstraint;
 use crate::service::choice::types::ChoiceService;
 use crate::service::voting::types::{VotingError, VotingService};
 use candid::Principal;
 use shared::mvc::{HasRepository, Model, Repository};
+use shared::pageable::{Page, PageRequest};
 use shared::types::wallet::{VotingConfigId, VotingId};
 
 impl VotingService {
@@ -44,13 +44,15 @@ impl VotingService {
     }
 
     pub fn update_voting(
-        voting: &mut Voting,
+        id: &VotingId,
         new_name: Option<String>,
         new_description: Option<String>,
         new_winners_need: Option<usize>,
         editor: Principal,
         timestamp: u64,
     ) -> Result<(), VotingError> {
+        let mut voting = VotingService::get_voting(id)?;
+        
         // unwrapping because it should exist if it is listed
         let vc = VotingConfig::repo()
             .get(&voting.get_voting_config_id())
@@ -60,22 +62,28 @@ impl VotingService {
             VotingService::assert_winners_need_is_fine(&vc, *winners_need)?;
         }
 
-        VotingService::assert_editor_can_edit(&vc, editor, voting.get_proposer())?;
+        if !VotingService::editor_can_edit(&vc, editor, voting.get_proposer()) {
+            return Err(VotingError::EditorNotFoundInVotingConfig(editor));
+        }
 
         voting
             .update(new_name, new_description, new_winners_need, timestamp)
-            .map_err(VotingError::ValidationError)
+            .map_err(VotingError::ValidationError)?;
+        Voting::repo().save(voting);
+        
+        Ok(())
     }
 
     pub fn delete_voting(id: &VotingId, deleter: Principal) -> Result<(), VotingError> {
-        let voting = Voting::repo()
-            .get(id)
-            .ok_or(VotingError::VotingNotFound(*id))?;
+        let voting = VotingService::get_voting(id)?;
 
         let vc = VotingConfig::repo()
             .get(voting.get_voting_config_id())
             .unwrap();
-        VotingService::assert_editor_can_edit(&vc, deleter, voting.get_proposer())?;
+
+        if !VotingService::editor_can_edit(&vc, deleter, voting.get_proposer()) {
+            return Err(VotingError::EditorNotFoundInVotingConfig(deleter));
+        }
 
         let voting = Voting::repo().delete(id).unwrap();
 
@@ -97,5 +105,15 @@ impl VotingService {
         }
 
         Ok(())
+    }
+    
+    #[inline(always)]
+    pub fn get_voting(id: &VotingId) -> Result<Voting, VotingError> {
+        Voting::repo().get(id).ok_or(VotingError::VotingNotFound(*id))
+    }
+    
+    #[inline(always)]
+    pub fn list_votings(page_req: &PageRequest<(), ()>) -> Page<Voting> {
+        Voting::repo().list(page_req)
     }
 }
